server/requirements.txt
Flask
gunicorn
flask_cors
pyjokes
json
werkzeug.exceptions